refactor: add space speaker to custom hook
